Added test for successful TV details response

+++ src/main/java/com/cinemadice/tmdbapi/model/tv/LastEpisodeToAir.java
@@ -22,7 +22,7 @@ public class LastEpisodeToAir {
     @SerializedName("still_path")
     private String stillPath;
     @SerializedName("vote_average")
-    protected double voteAverage;
+    protected float voteAverage;
     @SerializedName("vote_count")
     protected int voteCount;
